Remove all events, coloring, less destroy
Added the ability to remove all the events of a repeating event instead
of just one occurrence.  I also changed the coloring on myCalendar away
from red and removed some of the other destroys to prevent adding an
event from breaking things.  I also changed the Spring Semester to work
on 2015 instead of 2016 so that it make a better example.